GA-10 Changes updated by GEN AI tool.
