+++ cypress/e2e/API_TESTING/pets/findPetsByStatus.cy.js
@@ -8,7 +8,39 @@ requestInfo.url = apiBaseURL + requestInfo.url;
 
 describe('Finds Pets by status', () => {
     it('Multiple status values can be provided with comma separated strings', () => {
-        cy.fixture('200_application_json__findPetsByStatus').then(
+        cy.fixture('400_application_xml__findPetsByStatus').then(
+            (fixtureResponse) => {
+                requestInfo.body = fixtureResponse.payload
+                    ? fixtureResponse.payload
+                    : '';
+                requestInfo.headers = fixtureResponse.headers
+                    ? fixtureResponse.headers
+                    : '';
+
+                requestInfo.qs = fixtureResponse.queryParam
+                    ? fixtureResponse.queryParam
+                    : '';
+
+                cy.request(requestInfo).then((response) => {
+                    expect(response.status).to.eq(
+                        parseInt(fixtureResponse.responseStatusCode)
+                    );
+                    if (
+                        fixtureResponse.responseSchema &&
+                        fixtureResponse.responseSchema != ''
+                    ) {
+                        const validate = ajv.compile(
+                            fixtureResponse.responseSchema
+                        );
+                        const isValid = validate(response.body);
+                        expect(isValid).to.be.true;
+                    }
+                });
+            }
+        );
+    });
+    it('Multiple status values can be provided with comma separated strings', () => {
+        cy.fixture('403_application_json__findPetsByStatus').then(
             (fixtureResponse) => {
                 requestInfo.body = fixtureResponse.payload
                     ? fixtureResponse.payload
