+++ cypress/fixtures/200_application_json__update.json
@@ -1,0 +1,16 @@
+{
+    "headers": {
+        "accept": "application/json",
+        "header": ""
+    },
+    "responseStatusCode": "200",
+    "responseValue": {
+        "code": 201,
+        "type": "unknown",
+        "message": "updated succesfully"
+    },
+    "responseSchema": "",
+    "pathParam": {
+        "undefined": ""
+    }
+}
