+++ cypress/fixtures/403_application_json__findPetsByStatus.json
@@ -1,0 +1,25 @@
+{
+    "headers": {
+        "accept": "application/json",
+        "header": ""
+    },
+    "responseStatusCode": "403",
+    "responseSchema": {
+        "type": "object",
+        "properties": {
+            "code": {
+                "type": "integer",
+                "format": "int32"
+            },
+            "type": {
+                "type": "string"
+            },
+            "message": {
+                "type": "string"
+            }
+        }
+    },
+    "queryParam": {
+        "undefined": ""
+    }
+}
